The development of the user information modification function for the mini program has been completed, and the front-end administrator review function has also been completed

+++ src/main/java/com/fangxiong/mapper/GoodsMapper.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 @Mapper
 public interface GoodsMapper {
-    Integer selectCount();
+    Long selectCount();
     List<RecentGoodsVO> selectRecentGoods();
     List<TodoVO> selectWaitCheck();
     List<Goods> findAuditGoods();
